# Make a local copy of CUB's CachingDeviceAllocator and patch it for the race condition