As discussed in #237.