@fwyzard 2019-01-08
No changes in physics performance (as expected):

RelValTTbar_13 | reference-10824.5 | development-10824.5 | development-10824.8 | testing-10824.8
-- | -- | -- | -- | --
Efficiency | 0.4818 | 0.4824 | 0.4842 | 0.4842
Number of TrackingParticles (after cuts) | 5556 | 5556 | 5556 | 5556
Number of matched TrackingParticles | 2677 | 2680 | 2690 | 2690
Fake rate | 0.0519 | 0.0517 | 0.0337 | 0.0337
Duplicate rate | 0.0168 | 0.0175 | 0.0176 | 0.0176
Number of tracks | 32452 | 32480 | 31796 | 31796
Number of true tracks | 30769 | 30801 | 30723 | 30724
Number of fake tracks | 1683 | 1679 | 1073 | 1072
Number of pileup tracks | 27093 | 27118 | 26998 | 26998
Number of duplicate tracks | 546 | 567 | 559 | 559