@fwyzard 2019-01-08
No changes in throughput:
#### before
```
Warming up
Running 4 times over 4200 events with 1 jobs, each with 8 threads, 8 streams and 1 GPUs
  1730.3 ±   1.8 ev/s (4000 events)
  1726.9 ±   2.4 ev/s (4000 events)
  1731.3 ±   1.4 ev/s (4000 events)
  1738.8 ±   1.3 ev/s (4000 events)
```
#### after
```
Warming up
Running 4 times over 4200 events with 1 jobs, each with 8 threads, 8 streams and 1 GPUs
  1726.4 ±   2.6 ev/s (4000 events)
  1734.6 ±   1.8 ev/s (4000 events)
  1725.1 ±   1.8 ev/s (4000 events)
  1729.6 ±   2.0 ev/s (4000 events)
```