@fwyzard 2019-01-08
Given the simple commit history, I have tried merging this PR using the "Rebase and merge" approach instead of the usual "Squash and merge".

The positive aspect is that it keeps the full git log history, i.e. it looked like this:
* 5579cacfa29a Move mutex unlock after the cudaEventRecord() in DeviceFree()
* 9781dcf1d423 Include local copy of CachingDeviceAllocator
* eae2fbc643e3 Set the data pointer of error's SimpleVector (#236)
* ebc280c9d24e Fix a race for CUDA event in CachingHostAllocator (#237)
* ...

The negative aspect is that there is no trace of the PR number in the git log.
As part of the test, I have fixed that locally and force-pushed the update, so the history now looks like this:
* 68f320f5122e Move mutex unlock after the cudaEventRecord() in DeviceFree() (#240)
* 096e0e3aa83b Include local copy of CachingDeviceAllocator (#240)
* eae2fbc643e3 Set the data pointer of error's SimpleVector (#236)
* ebc280c9d24e Fix a race for CUDA event in CachingHostAllocator (#237)
* ...

What do people like best ?
  - squash into a single commit (I would still do this for very long commit histories)
  - rebase
  - rebase and reword to add the PR number
  - merge (as done in the upstream CMSSW)