@rovere 2019-01-08
Ciao @fwyzard 
I'm not sure I understand the difference between the rebase+reword and the usual merge approach.
Having the PR is, IMHO, very useful. I'd actually go one step further and add the username of the requester together with the PR number (also for our upstream CMSSW repository).