@makortel 2019-01-08
I'd prefer squash or rebase over the merge (since we're not the upstream repo, and "random" merges tend to make the history, and later comparisons/ports to upstream more difficult). I also think the PR number in the commit message(s) do(es) have some value (especially since we have those already in the history). So I suppose my preference would be 
* rebase and reword when "it makes sense" to preserve the commit history of a PR
* otherwise squash