@fwyzard 2019-01-08
@makortel thanks for your comments, I think I agree with you.
Rebasing + rewording takes a bit of manual effort, so I'd reserve it for the cases where I think it makes most sense.